Separated contact form to its own component
